Replaced apperances of Device with Resource.
There was a lot of mixes between the two words when they where actually
referring to the same thing.
This will most likely break any preexisting test installations due
to some classes being renamed. But since nothing is runnign in full
production yet this is the time to make such a change.
The external device monitor has agreed to change the protocoll as well
to refer to resource instead of device for method names and parameters.

Also fixed a DLS bug in the test code...again.

Change-Id: I1751c5ae0bdd63746633a3ce02fe03b07510b6a2